requirements.txt
Flask
requests
openai>=1.0.0
python-dotenv
gunicorn==20.1.0
